requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.0
Django==3.0.3
django-celery-beat==2.0.0
django-cors-headers==3.2.1
django-fernet-fields==0.6
django-filter==2.2.0
django-model-utils==4.0.0
django-redis==4.12.1
django-timezone-field==4.0
djangorestframework==3.11.1
drf-yasg==1.17.1
elastic-apm==6.1.3
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.11
MarkupSafe==1.1.1
monotonic==1.5
mysqlclient==2.0.1
packaging==20.4
pandas==1.1.5
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-toolbelt==0.9.1
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
zipp==3.1.0
flower==0.9.7
tenacity==7.0.0
Unipath==1.1